Reduce binary size by not storing scores for algorithmically scored classes.
This changes rounding of scores for space-adjacent characters. Unfortunately,
it appears that Urdu for title length and Hungarian for article length were
dependent on the exact old roundings. This changeset contains an intervention
for Urdu to boost the space-adjacent scores for the Arabic script. This wasn't
enough to compensate for all rounding changes.